requirements.txt
matplotlib
xarray
netCDF4
cartopy
shapely
pyproj
pydap
seaborn
pydeck
requests>=2.32
pyecharts>=2.0.5
streamlit-echarts>=0.4.0
Pillow
kaggle
pandas
plotly
